Python/birth_chart/astro-nexus-backend/requirements.txt
fastapi
uvicorn
python-dotenv
pyswisseph==2.10.3.2
pytz
geopy
timezonefinder
pydantic
